# backend/internal/httpapi/openapi.yaml
openapi: 3.1.0
info:
  title: ScholarBuddy API
  version: 1.0.0
  description: Development API reference for ScholarBuddy.
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Health check
      responses: { '200': { description: Service is healthy } }
  /:
    get:
      summary: Welcome message
      responses: { '200': { description: Welcome message } }
  /api/v1/opportunities:
    get:
      summary: List opportunities
      parameters:
        - { name: type, in: query, schema: { type: string, enum: [all, scholarship, hackathon, internship, research, extras] } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: per_page, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses: { '200': { description: Paginated opportunities } }
    post:
      summary: Create an opportunity
      security: [{ supabaseBearer: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/OpportunityInput' } } } }
      responses: { '201': { description: Opportunity created }, '401': { description: Valid bearer token required } }
  /api/v1/opportunities/{id}:
    get:
      summary: Get an opportunity's complete details
      parameters: [{ $ref: '#/components/parameters/OpportunityID' }]
      responses: { '200': { description: Opportunity }, '404': { description: Not found } }
    patch:
      summary: Update an opportunity created by the authenticated user
      security: [{ supabaseBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/OpportunityID' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/OpportunityInput' } } } }
      responses: { '200': { description: Opportunity updated }, '403': { description: Not owner } }
  /api/v1/abbujaan/opportunities:
    get:
      summary: List pending opportunities for the admin dashboard
      security: [{ adminBearer: [] }]
      responses: { '200': { description: Paginated opportunities } }
  /api/v1/abbujaan/opportunities/{id}/approve:
    patch:
      summary: Approve a pending opportunity
      security: [{ adminBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/OpportunityID' }]
      responses: { '200': { description: Opportunity approved } }
  /api/v1/abbujaan/opportunities/{id}:
    delete:
      summary: Delete an opportunity
      security: [{ adminBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/OpportunityID' }]
      responses: { '204': { description: Opportunity deleted } }
  /api/v1/abbujaan/login:
    post:
      summary: Sign in to the admin dashboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses: { '200': { description: Admin bearer token }, '401': { description: Invalid credentials } }
components:
  securitySchemes:
    supabaseBearer: { type: http, scheme: bearer, bearerFormat: JWT, description: Supabase Google OAuth access token }
    adminBearer: { type: http, scheme: bearer, bearerFormat: JWT, description: Token from /api/v1/abbujaan/login }
  parameters:
    OpportunityID: { name: id, in: path, required: true, schema: { type: string, format: uuid }, description: Opportunity UUID }
  schemas:
    OpportunityInput:
      type: object
      required: [title, description, types]
      properties:
        title: { type: string, maxLength: 250 }
        description: { type: string, maxLength: 10000 }
        types:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string, enum: [scholarship, hackathon, internship, research, extras] }
        eligibility: { type: string }
        steps: { type: string }
        benefits: { type: string }
        link: { type: string, format: uri }
        referral: { type: string }